refactor sound headers
